pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-magic>=0.4.27",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "urllib3>=2.5.0",
    "werkzeug>=3.1.3",
]
